+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [IPA-6.0.0]
+- Update cluster default version to 1.23, 1.24 may also be upgraded to.
+- Use AWS add-on aws-ebs-csi-driver
 ## [1.0.14]
 ### Updates
 - Allow using either FSX or EFS